feat: rewrite CLI from Go to TypeScript/Bunli
Introduces new command groups (workspace, git) with top-level aliases
(new→workspace create, ls→workspace list, rm→workspace remove).
Migrates from Go implementation to TypeScript with Bunli framework.
